requirements.txt
motor==3.4.0
aiogram==3.7.0
python-dotenv==1.0.1
